Merge290tomaster

+++ core/src/main/java/com/sequenceiq/cloudbreak/service/java/JavaDefaultVersionCalculator.java
@@ -21,6 +21,6 @@ public class JavaDefaultVersionCalculator {
                     .get()
                     .getVersion();
         }
-        return javaVersion.intValue();
+        return javaVersion;
     }
 }
